requirements.txt
Flask==2.3.3
pandas==2.1.1
requests==2.31.0
python-dotenv==1.0.0
tweepy==4.14.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.103.0
openpyxl==3.1.2